requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
matplotlib-inline==0.1.6
numpy==1.26.4
openai==1.29.0
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
playwright==1.41.2
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.0.1
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
transformers==4.40.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
